major BUGFIX: undefined behaviour caused by UseExtVoxels and IsDefault cluster members
since we are ot writing this members of Cluster to .root, they're not resetted in each iteration, and
propagated between tree entries. this caused undefined behaviour.

to solve that, VData got a ClearFilter() function that must be explicitly overrided when reading data classes
that present elements not written to disk. usually this only applies for filter tasks.

this function is not purely virtual but initializated to {} in the base clase